requirements.txt
attrs==19.3.0
Automat==0.8.0
blinker==1.6.3
boto3==1.28.29
botocore==1.31.29
certifi==2019.11.28
cftime==1.6.2
chardet==3.0.4
click==8.1.7
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cssselect==1.2.0
cycler==0.12.1
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
faiss-gpu==1.7.2
flask==3.0.0
fonttools==4.43.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.3
more-itertools==4.2.0
netCDF4==1.6.4
netifaces==0.10.4
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
pexpect==4.6.0
Pillow==10.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.1
pyquery==2.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==5.3.1
regex==2023.10.3
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
s3transfer==0.6.2
scikit-learn==1.3.1
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0.post10
sos==4.1
ssh-import-id==5.10
systemd-python==234
textblob==0.17.1
threadpoolctl==3.2.0
tqdm==4.66.1
tweepy==4.14.0
Twisted==18.9.0
tzdata==2023.3
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
werkzeug==3.0.0
zipp==3.17.0
zope.interface==4.7.1
